Made everything plugin compatible
- Made renderer backend a plugin
- Made the first attempt at making game logic a plugin
- Moved stuff to make it more dynamic library loader compatible
- Fixed getting default allocators by using CRTP

+++ engine/src/renderer/renderer_frontend.h
@@ -34,7 +34,7 @@ namespace C3D
 		bool Init(const RenderSystemConfig& config) override;
 		void Shutdown() override;
 
-		void OnResize(u16 width, u16 height);
+		void OnResize(u32 width, u32 height);
 
 		bool DrawFrame(RenderPacket* packet);
